cardApp.py
This works and is okay. However it's easy to break if a user doesn't input exactly what is asked for. Next step would be to figure out a way to check the input for Values 2-10, or A, K, Q or J (upper case necessary) and for the Suit input they need to put H, C, S or D.